requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.9.0
astroid==3.2.4
Authlib==1.5.2
babel==2.17.0
beautifulsoup4==4.10.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
configparser==7.2.0
contourpy==1.3.3
coverage==7.13.0
cryptography==45.0.2
cxxfilt==0.3.0
cycler==0.12.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.16
docutils==0.19
email-validator==2.3.0
fastapi==0.115.12
filelock==3.18.0
flake8==7.3.0
fonttools==4.61.1
fsspec==2025.5.0
gitdb==4.0.12
GitPython==3.1.43
google-adk==0.5.0
google-api-core==2.25.0rc1
google-api-python-client==2.157.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.91.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-logging==3.11.2
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.18.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
iniconfig==2.3.0
isort==5.13.2
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.9
librt==0.7.4
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mccabe==0.7.0
mcp==1.9.0
mypy==1.19.1
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.6
openai==1.72.0
opentelemetry-api==1.33.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==25.0
pandas==2.2.2
pathlib2==2.3.7.post1
pathspec==0.12.1
pep8==1.7.1
pillow==12.0.0
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.2.5
pyparsing==3.2.3
pyright==1.1.345
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
rust-demangler==1.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.2.1
Sphinx==6.0.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sse-starlette==2.3.5
starlette==0.46.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
tree-sitter==0.23.2
tree-sitter-cpp==0.23.4
tree-sitter-go==0.23.4
tree-sitter-java==0.23.5
tree-sitter-languages==1.10.2
tree-sitter-python==0.23.6
tree-sitter-rust==0.23.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
yapf==0.40.1
zipp==3.21.0
